Исправляет замечания по дублированию кода в статистике
